Rename app to App
admin to Admin
johari to Johari
cohort to Cohort
main to Main
login to Login
user to User
