+++ src/App/Sidebar/Sidebar.test.js
